On Mysql, unsigned smalint can cover till 65536. Seems that on MS SQL does not support this range (instead, supports only till 32768). Also, due to something i do not understand, cannot change the type of "level" field because it gas a constraint. So, as long as int and smallint are compatibles at query leve, we have changed it on initial creation to "PsotiveInteger" and shoudl be fine. (The already created database will not suffer from this change)

+++ server/src/uds/migrations/0001_initial.py
@@ -131,7 +131,7 @@ class Migration(migrations.Migration):
                 ('owner_type', models.SmallIntegerField(default=0, db_index=True)),
                 ('created', models.DateTimeField(db_index=True)),
                 ('source', models.CharField(default='internal', max_length=16, db_index=True)),
-                ('level', models.PositiveSmallIntegerField(default=0, db_index=True)),
+                ('level', models.PositiveIntegerField(default=0, db_index=True)),
                 ('data', models.CharField(default='', max_length=255)),
             ],
             options={
